sdk-core: bump to v14, migrate provider stack
- sdk-core 13.17.2 → 14.2.9; native-auth-server pinned 2.0.0
- drop @multiversx/sdk-network-providers (merged into sdk-core)
- rewrite GenericAbiService on SmartContractQuery + ArgSerializer;
  public getGenericData(interaction) shape preserved
- fix v14 signature shifts in task-runner, router, mx.proxy.service

+++ src/modules/router/services/router.transactions.service.ts
@@ -386,8 +386,10 @@ export class RouterTransactionService {
                     }),
                 )
                 .withGasLimit(
-                    args.addressRoute.length *
-                        gasConfig.router.multiPairSwapMultiplier,
+                    BigInt(
+                        args.addressRoute.length *
+                            gasConfig.router.multiPairSwapMultiplier,
+                    ),
                 )
                 .withChainID(mxConfig.chainID)
                 .buildTransaction()
